Moved main.go to cmd and moved formatter and transformer to metric
