+++ .github/workflows/go-lint.yaml
@@ -24,7 +24,7 @@ permissions:
 
 env:
   GO_VERSION: 1.23
-  GOLANGCI_LINT_VERSION: v2.1.2
+  GOLANGCI_LINT_VERSION: v2.9.0
 
 jobs:
   # Initial Setup Job
