+++ go/ironic-hardware-exporter/Dockerfile
@@ -14,7 +14,7 @@ COPY internal/ internal/
 # was called. For example, if we call make docker-build in a local env which has the Apple Silicon M1 SO
 # the docker BUILDPLATFORM arg will be linux/arm64 when for Apple x86 it will be linux/amd64. Therefore,
 # by leaving it empty we can ensure that the container and binary shipped on it will have the same platform.
-RUN CGO_ENABLED=0 GOOS=${TARGETOS:-linux} GOARCH=${TARGETARCH} go build -a -o ironic-hardware-exporter cmd/ironic-hardware-exporter/main.go
+RUN CGO_ENABLED=0 GOOS=${TARGETOS:-linux} GOARCH=${TARGETARCH} go build -a -o ironic-hardware-exporter cmd/main.go
 
 FROM gcr.io/distroless/static:nonroot
 WORKDIR /
