+++ go/ironic-hardware-exporter/internal/metrics/formatter_test.go
@@ -1,4 +1,4 @@
-package server
+package metrics
 
 import (
 	"strings"
