+++ go/ironic-hardware-exporter/internal/metrics/transformer_test.go
@@ -1,4 +1,4 @@
-package server
+package metrics
 
 import (
 	"testing"
