+++ go/ironic-hardware-exporter/local/rabbitmq/.gitignore
@@ -1,0 +1,4 @@
+certs/*
+!certs/.gitkeep
+logs/*
+!logs/.gitkeep
