+++ go/ironic-hardware-exporter/local/rabbitmq/config/enabled_plugins
@@ -1,0 +1,1 @@
+[rabbitmq_management].
