+++ go/ironic-hardware-exporter/local/rabbitmq/config/rabbitmq.conf
@@ -1,0 +1,11 @@
+listeners.tcp.default = 5672
+management.tcp.port = 15672
+
+listeners.ssl.default = 5671
+ssl_options.cacertfile = /etc/rabbitmq/tls/ca.pem
+ssl_options.certfile = /etc/rabbitmq/tls/server.pem
+ssl_options.keyfile = /etc/rabbitmq/tls/server-key.pem
+ssl_options.verify = verify_none
+ssl_options.fail_if_no_peer_cert = false
+ssl_options.versions.1 = tlsv1.2
+ssl_options.versions.2 = tlsv1.3
